Improved Segmentation Methods
The previous segmentation methods used distance transformation uniformly across the dataset, which assume the distribution of skeleton nodes are also uniformly. However, when facing skeletons generated from vessels with more diverse range of radius, it is better for the radius around a skeleton node to be considered when performing distance transforming.