Added pre-processing step
This step relabels part of small branches that are close to thicker branches. This prevents part of the "dent problem" in visualization.
The dent problem is caused by, at branching points, small branches does not shrink immediately, meaning that at the connection between a big branch and a small branch, the initial nodes' size are not too different. This causes problems in the distance transform calculations, because a large ellipsoid would be labeled as a part of the small branch, creating a dent.
This pre-processing step has no statistical related importance, and is added purely for better visualization.

Co-Authored-By: gnuoyea <55409401+gnuoyea@users.noreply.github.com>